scripts/requirements.txt
cloudera-director-python-client==2.1.1
pyhocon==0.3.8